Moved api functions to separate controller files
